Moved KeyCodeConverter into the Json.Converters namespace.
